Change endpoint and controller to "events" instead of "webhooks" and return a default set of events for now, for testing purposes.
