+++ implementation/sem-webhook-service/RestService/Events/EventDto.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace RestService.Events
+{
+    public record EventDto(Guid Id, string Name);
+}
