--- implementation/sem-webhook-service/RestService/Webhooks/Webhook.cs
@@ -1,6 +1,0 @@
-﻿namespace RestService.Webhooks
-{
-    public class Webhook
-    {
-    }
-}
